submodule: remove the afamqp support
Reason:
The afamqp needs the rabbitmq-c as a submodule, but
the rabbitmq-c has a patch (3cb254d902a9b226bf95696af3a98839bb7797a4)
which is not valid, therefore we could not push it to our gitlab server.

Note: The problem is not at the gitlab site, this patch is not push-able to the github too.

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>