tgz2build: ZBS integration
tgz2build: use local stylesheet for compiling manpages

tgz2build: list required files explicitly (omitting libtest)

If it is not specified, the default is an internet resource from
docbook.sourceforge.net, which can be broken. If the `docbook-xsl`
package is installed, it is used instead of downloading the files from
internet, but it is quite implicit. With hardcoding the file locations
we can detect if a builder machine does not have the required package.

Signed-off-by: Laszlo Budai <laszlo.budai@balabit.com>
Signed-off-by: Máté Farkas <mate.farkas@balabit.com>